feat: detect AMR-determining SNPs in promoters if specified
