+++ amira/construct_graph.py
@@ -1,16 +1,11 @@
-import json
 import os
-import shutil
 import statistics
-import subprocess
 import sys
 from collections import Counter, defaultdict, deque
-from itertools import combinations, product
+from itertools import product
 from multiprocessing import Pool
 
 import numpy as np
-import pandas as pd
-import pysam
 import sourmash
 from joblib import Parallel, delayed
 from suffix_tree import Tree
@@ -2201,7 +2196,10 @@ class GeneMerGraph:
             for read in node.get_reads():
                 indices = [i for i, n in enumerate(self.get_readNodes()[read]) if n == node_hash]
                 positions = [self.get_readNodePositions()[read][i] for i in indices]
-                entire_read_sequence = fastq_data["_".join(read.split("_"))[:-1]]["sequence"]
+                if read.count("_") > 0:
+                    entire_read_sequence = fastq_data["_".join(read.split("_"))[:-1]]["sequence"]
+                else:
+                    entire_read_sequence = fastq_data[read]["sequence"]
                 for p in positions:
                     minhash.add_sequence(entire_read_sequence[p[0] : p[1] + 1], force=True)
             node_minhashes[node_hash] = minhash
@@ -2940,393 +2938,6 @@ class GeneMerGraph:
                             )
                     allele_counts[geneOfInterest] += 1
         return clustered_reads, cluster_copy_numbers, allele_counts
-
-    def get_closest_allele(self, bam_file_path, mapping_type, ref_cov_proportion=None):
-        valid_references = []
-        invalid_references = []
-        ref_covered = {}
-        ref_matching = {}
-        ref_lengths = {}
-        unique_reads = set()
-        # Open the BAM file
-        with pysam.AlignmentFile(bam_file_path, "rb") as bam_file:
-            for read in bam_file.fetch():
-                # check if the read is mapped
-                if read.is_unmapped:
-                    continue
-                # add the read name to a set
-                unique_reads.add(read.query_name)
-                # get the length of the reference
-                total_length = bam_file.get_reference_length(read.reference_name)
-                # add the reference to the coverage dictionary
-                if read.reference_name not in ref_covered:
-                    ref_covered[read.reference_name] = 0
-                    ref_matching[read.reference_name] = 0
-                    ref_lengths[read.reference_name] = total_length
-                # get the proportion of bases matching the reference
-                matching_bases = 0
-                for op, length in read.cigartuples:
-                    if op == 7:
-                        matching_bases += length
-                if mapping_type == "reads":
-                    prop_matching = (matching_bases / total_length) * 100
-                    prop_covered = ref_cov_proportion[read.reference_name]
-                if mapping_type == "allele":
-                    prop_matching = (matching_bases / read.infer_read_length()) * 100
-                    # get the proportion of the reference covered by the query
-                    prop_covered = read.query_alignment_length / total_length
-                # add the stats to the dictionaries
-                if prop_matching > ref_matching[read.reference_name]:
-                    ref_matching[read.reference_name] = prop_matching
-                if prop_covered > ref_covered[read.reference_name]:
-                    ref_covered[read.reference_name] = prop_covered
-        for ref in ref_matching:
-            if ref_covered[ref] >= 0.85:
-                valid_references.append(
-                    (ref, ref_matching[ref], ref_lengths[ref], ref_covered[ref])
-                )
-            else:
-                invalid_references.append(
-                    (ref, ref_matching[ref], ref_lengths[ref], ref_covered[ref])
-                )
-        valid_references = sorted(valid_references, key=lambda x: (x[1], x[2]), reverse=True)
-        if len(valid_references) != 0:
-            return True, valid_references, unique_reads
-        else:
-            invalid_references = sorted(
-                invalid_references, key=lambda x: (x[2], x[1]), reverse=True
-            )
-            return False, invalid_references, unique_reads
-
-    def create_output_dir(self, output_dir, allele_name):
-        output_dir = os.path.join(output_dir, allele_name)
-        if not os.path.exists(output_dir):
-            os.mkdir(output_dir)
-        return output_dir
-
-    def write_fasta(self, file_path, sequences):
-        with open(file_path, "w") as outFasta:
-            outFasta.write("\n".join(sequences))
-
-    def run_subprocess(self, command):
-        subprocess.run(command, shell=True, check=True)
-
-    def map_reads(
-        self, output_dir, reference_fasta, input_fasta, output_prefix, minimap_opts, minimap2_path
-    ):
-        sam_file = os.path.join(output_dir, f"{output_prefix}.mapped.sam")
-        bam_file = os.path.join(output_dir, f"{output_prefix}.mapped.bam")
-        map_command = f"{minimap2_path} {minimap_opts} {reference_fasta} {input_fasta} > {sam_file}"
-        self.run_subprocess(map_command)
-        sort_and_index_command = (
-            f"samtools sort {sam_file} > {bam_file} && samtools index {bam_file}"
-        )
-        self.run_subprocess(sort_and_index_command)
-        return bam_file
-
-    def racon_polish(
-        self,
-        output_dir,
-        racon_path,
-        input_fasta,
-        sam_file,
-        reference_fasta,
-        output_fasta,
-        window_length,
-    ):
-        racon_command = f"{racon_path} -u -t 1 --no-trimming -w {window_length} {input_fasta} "
-        racon_command += f"{sam_file} {reference_fasta} > {output_fasta}"
-        self.run_subprocess(racon_command)
-
-    def prepare_reference_sequences(self, gene_name, reference_genes, output_dir):
-        sequences = [f">{allele}\n{seq}" for allele, seq in reference_genes[gene_name].items()]
-        self.write_fasta(os.path.join(output_dir, "01.reference_alleles.fasta"), sequences)
-        return [len(seq) for seq in reference_genes[gene_name].values()]
-
-    def get_allele_sequence(self, gene_name, allele, reference_genes):
-        return reference_genes[gene_name][allele]
-
-    def racon_one_iteration(
-        self,
-        output_dir,
-        racon_path,
-        read_file,
-        sam_file,
-        sequence_to_polish,
-        polished_sequence,
-        window_size,
-        minimap2_path,
-    ):
-        # run minimap2
-        bam_file = self.map_reads(
-            output_dir,
-            os.path.join(output_dir, sequence_to_polish),
-            read_file,
-            sam_file.replace(".mapped.sam", ""),
-            "-a --MD -t 1 -x map-ont --eqx",
-            minimap2_path,
-        )
-        # run racon
-        self.racon_polish(
-            output_dir,
-            racon_path,
-            read_file,
-            os.path.join(output_dir, sam_file),
-            os.path.join(output_dir, sequence_to_polish),
-            os.path.join(output_dir, polished_sequence),
-            window_size,
-        )
-        # rename the polished sequence
-        shutil.copy(
-            os.path.join(output_dir, polished_sequence),
-            os.path.join(output_dir, sequence_to_polish),
-        )
-        return bam_file
-
-    def get_ref_allele_pileups(self, bam_file, output_dir):
-        read_depths = []
-        ref_allele_positions = {}
-        cov_proportion = {}
-        # Open the BAM file
-        with pysam.AlignmentFile(bam_file, "rb") as bam:
-            # Iterate over each reference (chromosome/contig)
-            for ref in bam.references:
-                ref_length = bam.get_reference_length(ref)
-                # Initialize a list to store read depths for each position
-                depth_list = [0] * ref_length
-                # Fetch all reads mapped to this reference
-                for read in bam.fetch(ref):
-                    if read.is_unmapped:
-                        continue
-                    # Iterate over the aligned positions of the read
-                    for pos in read.get_reference_positions():
-                        if pos < ref_length:
-                            depth_list[pos] += 1
-                # Convert the depth list to a string format
-                depth_list_str = ",".join(map(str, depth_list))
-                read_depths.append(f">{ref}\n{depth_list_str}")
-                # get the first and last non-zero element
-                try:
-                    first_index = depth_list.index(next(x for x in depth_list if x != 0))
-                    last_index = (
-                        len(depth_list)
-                        - 1
-                        - depth_list[::-1].index(next(x for x in reversed(depth_list) if x != 0))
-                    )
-                except StopIteration:
-                    first_index, last_index = None, None
-                ref_allele_positions[ref] = (first_index, last_index)
-                cov_proportion[ref] = len([d for d in depth_list if d != 0]) / len(depth_list)
-        # Write the read depths to the output file
-        output_file_path = os.path.join(output_dir, "reference_allele_coverages.txt")
-        with open(output_file_path, "w") as output_file:
-            output_file.write("\n".join(read_depths))
-        return ref_allele_positions, cov_proportion
-
-    def compare_reads_to_references(
-        self,
-        allele_file,
-        output_dir,
-        reference_genes,
-        racon_path,
-        fastqContent,
-        phenotypes,
-        debug,
-        minimap2_path,
-    ):
-        allele_name = os.path.basename(allele_file).replace(".fastq.gz", "")
-        gene_name = "_".join(allele_name.split("_")[:-1])
-        output_dir = self.create_output_dir(output_dir, allele_name)
-        self.prepare_reference_sequences(gene_name, reference_genes, output_dir)
-        bam_file = self.map_reads(
-            output_dir,
-            os.path.join(output_dir, "01.reference_alleles.fasta"),
-            allele_file,
-            "02.read",
-            "-a --MD -t 1 -x map-ont --eqx",
-            minimap2_path,
-        )
-        ref_allele_positions, ref_cov_proportion = self.get_ref_allele_pileups(bam_file, output_dir)
-        if debug is False:
-            os.remove(os.path.join(output_dir, "reference_allele_coverages.txt"))
-        validity, references, unique_reads = self.get_closest_allele(
-            bam_file, "reads", ref_cov_proportion
-        )
-        if validity is True:
-            valid_allele, match_proportion, match_length, coverage_proportion = references[0]
-            valid_allele_sequence = self.get_allele_sequence(
-                gene_name, valid_allele, reference_genes
-            )
-            first_base, last_base = ref_allele_positions[valid_allele]
-            self.write_fasta(
-                os.path.join(output_dir, "03.sequence_to_polish.fasta"),
-                [f">{valid_allele}\n{valid_allele_sequence[first_base: last_base+1]}"],
-            )
-            iterations = 5
-            for _ in range(iterations):
-                try:
-                    self.racon_one_iteration(
-                        output_dir,
-                        racon_path,
-                        allele_file,
-                        "02.read.mapped.sam",
-                        "03.sequence_to_polish.fasta",
-                        "04.polished_sequence.fasta",
-                        str(len(valid_allele_sequence)),
-                        minimap2_path,
-                    )
-                except subprocess.CalledProcessError:
-                    try:
-                        gene_name = valid_allele.split(".")[0]
-                        closest_ref = valid_allele.split(".")[1]
-                    except IndexError:
-                        gene_name = "_".join(allele_name.split("_")[:-1])
-                        closest_ref = valid_allele
-                    return {
-                        "Gene name": gene_name,
-                        "Sequence name": phenotypes[valid_allele],
-                        "Closest reference": closest_ref,
-                        "Reference length": match_length,
-                        "Identity (%)": 0,
-                        "Coverage (%)": 0,
-                        "Amira allele": allele_name,
-                        "Number of reads": len(unique_reads),
-                    }
-            bam_file = self.map_reads(
-                output_dir,
-                os.path.join(output_dir, "01.reference_alleles.fasta"),
-                os.path.join(output_dir, "04.polished_sequence.fasta"),
-                "05.read",
-                "-a --MD -t 1 --eqx",
-                minimap2_path,
-            )
-            validity, references, _ = self.get_closest_allele(bam_file, "allele")
-            max_similarity = references[0][1]
-            references = [r for r in references if r[1] == max_similarity]
-            if len(references) == 1:
-                closest_allele, match_proportion, match_length, coverage_proportion = references[0]
-                with open(os.path.join(output_dir, "04.polished_sequence.fasta")) as i:
-                    final_allele_sequence = "".join(i.read().split("\n")[1:])
-                self.write_fasta(
-                    os.path.join(output_dir, "06.final_sequence.fasta"),
-                    [f">{closest_allele}\n{final_allele_sequence}"],
-                )
-                try:
-                    gene_name = closest_allele.split(".")[0]
-                    closest_ref = closest_allele.split(".")[1]
-                except IndexError:
-                    gene_name = "_".join(allele_name.split("_")[:-1])
-                    closest_ref = closest_allele
-                return {
-                    "Gene name": gene_name,
-                    "Sequence name": phenotypes[closest_allele],
-                    "Closest reference": closest_ref,
-                    "Reference length": match_length,
-                    "Identity (%)": round(match_proportion, 1),
-                    "Coverage (%)": min(100.0, round(coverage_proportion * 100, 1)),
-                    "Amira allele": allele_name,
-                    "Number of reads": len(unique_reads),
-                }
-            if len(references) > 1:
-                closest_allele, match_proportion, match_length, coverage_proportion = [], [], [], []
-                for r in references:
-                    closest_allele.append(r[0])
-                    match_proportion.append(r[1])
-                    match_length.append(r[2])
-                    coverage_proportion.append(r[3])
-                with open(os.path.join(output_dir, "04.polished_sequence.fasta")) as i:
-                    final_allele_sequence = "".join(i.read().split("\n")[1:])
-                self.write_fasta(
-                    os.path.join(output_dir, "06.final_sequence.fasta"),
-                    [f">{'/'.join(closest_allele)}\n{final_allele_sequence}"],
-                )
-                try:
-                    gene_names = "/".join(
-                        sorted(list(set([c.split(".")[0] for c in closest_allele])))
-                    )
-                    closest_refs = "/".join([c.split(".")[1] for c in closest_allele])
-                    phenotypes = "/".join([phenotypes[c] for c in closest_allele])
-                except IndexError:
-                    gene_names = "_".join(allele_name.split("_")[:-1])
-                    closest_refs = "/".join(closest_allele)
-                    phenotypes = "/".join([phenotypes[c] for c in closest_allele])
-                return {
-                    "Gene name": gene_names,
-                    "Sequence name": phenotypes,
-                    "Closest reference": closest_refs,
-                    "Reference length": "/".join([str(m) for m in match_length]),
-                    "Identity (%)": "/".join([str(round(p, 1)) for p in match_proportion]),
-                    "Coverage (%)": "/".join(
-                        [str(min(100.0, round(p * 100, 1))) for p in coverage_proportion]
-                    ),
-                    "Amira allele": allele_name,
-                    "Number of reads": len(unique_reads),
-                }
-        else:
-            if len(references) != 0:
-                invalid_allele, match_proportion, match_length, coverage_proportion = references[0]
-                try:
-                    gene_name = invalid_allele.split(".")[0]
-                    closest_ref = invalid_allele.split(".")[1]
-                except IndexError:
-                    gene_name = "_".join(allele_name.split("_")[:-1])
-                    closest_ref = invalid_allele
-                return {
-                    "Gene name": gene_name,
-                    "Sequence name": phenotypes[invalid_allele],
-                    "Closest reference": closest_ref,
-                    "Reference length": match_length,
-                    "Identity (%)": round(match_proportion, 1),
-                    "Coverage (%)": min(100.0, round(coverage_proportion * 100, 1)),
-                    "Amira allele": allele_name,
-                    "Number of reads": len(unique_reads),
-                }
-            else:
-                return {
-                    "Gene name": "",
-                    "Sequence name": "",
-                    "Closest reference": "",
-                    "Reference length": 0,
-                    "Identity (%)": 0,
-                    "Coverage (%)": 0,
-                    "Amira allele": allele_name,
-                    "Number of reads": len(unique_reads),
-                }
-
-    def get_alleles(
-        self,
-        readFiles,
-        racon_path,
-        threads,
-        output_dir,
-        reference_genes,
-        fastqContent,
-        phenotypes_path,
-        debug,
-        minimap2_path,
-    ):
-        # import the phenotypes
-        with open(phenotypes_path) as i:
-            phenotypes = json.load(i)
-        # batch the read files for multi processing
-        job_list = [readFiles[i : i + threads] for i in range(0, len(readFiles), threads)]
-        gene_data = []
-        for subset in tqdm(job_list):
-            gene_data += Parallel(n_jobs=threads)(
-                delayed(self.compare_reads_to_references)(
-                    r,
-                    output_dir,
-                    reference_genes,
-                    racon_path,
-                    fastqContent.get("_".join(os.path.basename(r).split("_")[:-1]), None),
-                    phenotypes,
-                    debug,
-                    minimap2_path,
-                )
-                for r in subset
-            )
-        gene_data = sorted(gene_data, key=lambda x: x["Gene name"])
-        return pd.DataFrame(gene_data)
 
     def remove_non_AMR_associated_nodes(self, genesOfInterest):
         # get the reads containing AMR genes
